refactor: address CodeRabbit feedback & fix lint/type errors
