--- test_output.txt
@@ -1,44 +1,0 @@
-FAIL test/ui/a2ui-renderer.test.ts
-  A2UIRenderer
-    Component Rendering
-      ÔêÜ should render a text component (72 ms)
-      ├ù should render a container with children (19 ms)
-      ÔêÜ should render a button component (21 ms)
-      ÔêÜ should handle unknown component type gracefully (26 ms)
-    Style Sanitization
-      ÔêÜ should apply allowed styles (31 ms)
-      ÔêÜ should block javascript: URLs in styles (14 ms)
-      ÔêÜ should block url() in styles (15 ms)
-      ÔêÜ should block expression() in styles (IE attack vector) (16 ms)
-    Button Actions
-      ÔêÜ should dispatch a2ui-action event on button click (16 ms)
-      ÔêÜ should not set up onClick for buttons without action (15 ms)
-    Switch Component
-      ÔêÜ should handle checked = false correctly (15 ms)
-      ÔêÜ should handle checked = undefined correctly (15 ms)
-    Error Handling
-      ÔêÜ should display error div on render failure (15 ms)
-
-  ÔùÅ A2UIRenderer ÔÇ║ Component Rendering ÔÇ║ should render a container with children
-
-    expect(received).toBe(expected) // Object.is equality
-
-    Expected: "Child 1"
-    Received: ""
-
-      209 | 			const children = containerEl.querySelectorAll('.a2ui-text');
-      210 | 			expect(children.length).toBe(2);
-    > 211 | 			expect(children[0].textContent).toBe('Child 1');
-          | 			                                ^
-      212 | 			expect(children[1].textContent).toBe('Child 2');
-      213 | 			expect(MarkdownRenderer.render).toHaveBeenCalledTimes(2);
-      214 | 		});
-
-      at test/ui/a2ui-renderer.test.ts:211:36
-      at fulfilled (node_modules/tslib/tslib.js:167:62)
-
-Test Suites: 1 failed, 1 total
-Tests:       1 failed, 12 passed, 13 total
-Snapshots:   0 total
-Time:        1.702 s
-Ran all test suites.
